Remove need for toplevel funding controller
